Added basic testing for cot-testcase cmd
