+++ gen3-cli/tests/data/system/testcases/testcase-failed.yml
@@ -1,5 +1,5 @@
 ---
-  testThree:
+  testFailed:
     setup:
       -
         cmd: echo "Hello $NAME"
